fix: fixes #2880
