+++ blog/views/__init__.py
@@ -1,0 +1,1 @@
+from .post import PostListView, PostDetailView, PostCreateView, PostUpdateView, PostDeleteView
