+++ comments/views/__init__.py
@@ -1,0 +1,1 @@
+from .comment import add_comment, delete_comment, edit_comment, reply_comment, autocomplete
