+++ company/views/__init__.py
@@ -1,0 +1,32 @@
+from .company_dashboard import (
+    RegisterCompanyView, 
+    CompanyDashboardAnalyticsView, 
+    CompanyDashboardManageBugsView, 
+    CompanyDashboardManageDomainsView, 
+    CompanyDashboardManageRolesView, 
+    CompanyDashboardManageBughuntView,
+    company_view
+    )
+
+from .helper import (
+    validate_company_user,
+    check_company_or_manager,
+    delete_prize,
+    edit_prize,
+    delete_manager,
+    accept_bug,
+)
+
+from .domain import (
+    AddDomainView,
+    DomainView,
+)
+
+from .issue import (
+    ShowBughuntView,
+    EndBughuntView,
+    AddHuntView,
+)
+
+
+
